Implemented document controller

+++ src/main/java/local/project/api/model/DocumentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentRepository extends CrudRepository<DocumentEntity, Integer> {
+public interface DocumentRepository extends CrudRepository<DocumentEntity, Long> {
 }
